Trying to improve how the website components responds to eachother to avoid jumping and increase userX. Moved buttons together and added backwardsbutton and put visible=hidden on objects that would be rendered later.
